make replace_underscore private, move to attr/field.rs
